+++ OdhApiCore/Controllers/api/OdhTagController.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
-using OdhApiCore.Controllers.helper;
+using OdhApiCore.Filters;
 using SqlKata.Execution;
 using System;
 using System.Collections.Generic;
