+++ OdhApiCore/Controllers/other/LicenseCountController.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using OdhApiCore.Controllers.helper;
+using OdhApiCore.Filters;
 using SqlKata.Execution;
 using System;
 using System.Collections.Generic;
